requirements.txt
beautifulsoup4>=4.12.3
progressbar2>=4.2.0
Requests>=2.32.2
retry>=0.9.2
